Add tests for SiteNav, remove obsolete tests for Header
