+++ frontend/src/components/SiteNav.js
@@ -46,34 +46,36 @@ const SiteNav = ({
 
   return (
     <div className="smart-hub-sitenav position-relative font-ui text-white smart-hub-bg-blue smart-hub-height-hd minh-tablet no-print">
-      <div className="width-full">
-        <div className="margin-x-2 border-bottom smart-hub-border-darkblue padding-y-2">
-          <p className="text-bold margin-y-105">{ user.name }</p>
-          <p className="font-sans-3xs margin-y-105">{ user.email }</p>
-        </div>
-      </div>
-      <nav className="width-full margin-bottom-2">
-        <ul className="add-list-reset">
-          <li>
-            <Link
-              className="display-block border-left-05 border-white text-white text-bold text-no-underline padding-2 margin-y-2"
-              to="/activity-reports"
-            >
-              <span className="padding-right-105">
-                <FontAwesomeIcon color="white" icon={faChartBar} />
-              </span>
-              Activity Reports
-            </Link>
-          </li>
-        </ul>
-      </nav>
-      <div className="width-full position-absolute bottom-0 padding-bottom-6">
-        {authenticated && (
-          <ul className="add-list-reset margin-x-2 padding-top-2 text-base-lighter border-top smart-hub-border-darkblue">
-            {items.map((item, i) => (<li key={`smart-hub-nav__item-${i}`}>{item}</li>))}
-          </ul>
-        )}
-      </div>
+      {authenticated && (
+        <>
+          <div className="width-full">
+            <div className="margin-x-2 border-bottom smart-hub-border-darkblue padding-y-2">
+              <p className="text-bold margin-y-105">{ user.name }</p>
+              <p className="font-sans-3xs margin-y-105">{ user.email }</p>
+            </div>
+          </div>
+          <nav className="width-full margin-bottom-2">
+            <ul className="add-list-reset">
+              <li>
+                <Link
+                  className="display-block border-left-05 border-white text-white text-bold text-no-underline padding-2 margin-y-2"
+                  to="/activity-reports"
+                >
+                  <span className="padding-right-105">
+                    <FontAwesomeIcon color="white" icon={faChartBar} />
+                  </span>
+                  Activity Reports
+                </Link>
+              </li>
+            </ul>
+          </nav>
+          <div className="width-full position-absolute bottom-0 padding-bottom-6">
+            <ul className="add-list-reset margin-x-2 padding-top-2 text-base-lighter border-top smart-hub-border-darkblue">
+              {items.map((item, i) => (<li key={`smart-hub-nav__item-${i}`}>{item}</li>))}
+            </ul>
+          </div>
+        </>
+      )}
     </div>
   );
 };
@@ -82,13 +84,14 @@ SiteNav.displayName = 'SiteNav';
 SiteNav.propTypes = {
   admin: PropTypes.bool,
   authenticated: PropTypes.bool,
-  logout: PropTypes.func.isRequired,
+  logout: PropTypes.func,
   user: PropTypes.shape({ name: PropTypes.string, email: PropTypes.string }),
 };
 
 SiteNav.defaultProps = {
   admin: false,
   authenticated: false,
+  logout: () => {},
   user: {
     name: '',
     email: '',
